+++ buildSrc/src/main/kotlin/Plugins.kt
@@ -14,18 +14,27 @@
  * limitations under the License.
  */
 
-package deps
-
 object Plugins {
+
+  object BuildPlugins {
+    const val androidLib = "com.android.library"
+    const val application = "com.android.application"
+    const val kotlinAndroid = "kotlin-android"
+    const val kotlinKapt = "kotlin-kapt"
+    const val mavenPublish = "maven-publish"
+    const val navSafeArgs = "androidx.navigation.safeargs.kotlin"
+    const val spotless = "com.diffplug.spotless"
+  }
+
+  // classpath plugins
   const val androidGradlePlugin = "com.android.tools.build:gradle:${Versions.androidGradlePlugin}"
   const val kotlinGradlePlugin = "org.jetbrains.kotlin:kotlin-gradle-plugin:${Dependencies.Versions.Kotlin.stdlib}"
   const val navSafeArgsGradlePlugin = "androidx.navigation:navigation-safe-args-gradle-plugin:${Dependencies.Versions.Androidx.navigation}"
   const val spotlessGradlePlugin = "com.diffplug.spotless:spotless-plugin-gradle:${Versions.spotlessGradlePlugin}"
 
   object Versions {
-    const val androidGradlePlugin = "4.2.0-beta02"
+    const val androidGradlePlugin = "4.2.0-beta06"
     const val buildTools = "30.0.2"
     const val spotlessGradlePlugin = "5.11.0"
   }
-
 }
